Rename inner editor package to deck
