+++ src/main/scala/editor/gui/ccp/data/CategoryTransferData.scala
@@ -2,7 +2,7 @@ package editor.gui.ccp.data
 
 import editor.collection.Categorization
 import editor.database.card.Card
-import editor.gui.editor.EditorFrame
+import editor.gui.deck.EditorFrame
 
 import java.awt.datatransfer.DataFlavor
