+++ src/main/scala/editor/gui/deck/CalculateHandPanel.scala
@@ -1,4 +1,4 @@
-package editor.gui.editor
+package editor.gui.deck
 
 import editor.collection.mutable.Deck
 import editor.gui.settings.SettingsDialog
